requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
torch==2.1.0
torchvision==0.16.0
open-clip-torch==2.20.0
ultralytics==8.0.20
facenet-pytorch==2.5.3
pillow==10.0.0
pyyaml==6.0.1
python-multipart==0.0.6
tqdm==4.66.1